seperate bindings layer
