+++ mina-transaction/gen/v1/dune
@@ -1,0 +1,16 @@
+(rule
+ (targets
+  js-layout.ts
+  transaction.ts
+  transaction-json.ts
+  transaction-bigint.ts)
+ (mode promote)
+ (enabled_if
+  (<> %{env:NO_JS_BUILD=0} 1))
+ (deps
+  (:script ../../../../build/js-layout-to-types.mjs)
+  ../../../ocaml/jsLayout.json
+  ../../../../../.prettierrc.cjs
+  (glob_files ../../../../../node_modules/prettier/*))
+ (action
+  (run node %{script})))
