+++ mina-transaction/gen/v1/transaction.ts
@@ -20,16 +20,17 @@ import {
   MayUseToken,
   VerificationKeyHash,
   ReceiptChainHash,
-} from '../v1/transaction-leaves-bigint.js';
-import { GenericSignable } from '../../lib/generic.js';
-import { SignableFromLayout, GenericLayout } from '../../lib/from-layout.js';
+} from '../../v1/transaction-leaves.js';
+import { GenericProvableExtended } from '../../../lib/generic.js';
+import { ProvableFromLayout, GenericLayout } from '../../../lib/from-layout.js';
 import * as Json from './transaction-json.js';
+import * as Value from './transaction-bigint.js';
 import { jsLayout } from './js-layout.js';
 
 export { customTypes, ZkappCommand, AccountUpdate, Account };
 export { Json };
-export * from '../v1/transaction-leaves-bigint.js';
-export { signableFromLayout, toJSONEssential, empty, Layout, TypeMap };
+export * from '../../v1/transaction-leaves.js';
+export { provableFromLayout, toJSONEssential, empty, Layout, TypeMap };
 
 type TypeMap = {
   PublicKey: PublicKey;
@@ -44,7 +45,7 @@ type TypeMap = {
 };
 
 const TypeMap: {
-  [K in keyof TypeMap]: Signable<TypeMap[K], Json.TypeMap[K]>;
+  [K in keyof TypeMap]: ProvableExtended<TypeMap[K], Value.TypeMap[K], Json.TypeMap[K]>;
 } = {
   PublicKey,
   UInt64,
@@ -57,60 +58,81 @@ const TypeMap: {
   Bool,
 };
 
-type Signable<T, TJson> = GenericSignable<T, TJson, Field>;
+type ProvableExtended<T, TValue, TJson> = GenericProvableExtended<T, TValue, TJson, Field>;
 type Layout = GenericLayout<TypeMap>;
 
 type CustomTypes = {
-  TransactionVersion: Signable<UInt32, Json.TypeMap['UInt32']>;
-  ZkappUri: Signable<
+  TransactionVersion: ProvableExtended<UInt32, Value.TypeMap['UInt32'], Json.TypeMap['UInt32']>;
+  ZkappUri: ProvableExtended<
     {
       data: string;
       hash: Field;
     },
+    {
+      data: string;
+      hash: Value.TypeMap['Field'];
+    },
     string
   >;
-  TokenSymbol: Signable<
+  TokenSymbol: ProvableExtended<
     {
       symbol: string;
       field: Field;
     },
+    {
+      symbol: string;
+      field: Value.TypeMap['Field'];
+    },
     string
   >;
-  StateHash: Signable<Field, Json.TypeMap['Field']>;
-  BalanceChange: Signable<
+  StateHash: ProvableExtended<Field, Value.TypeMap['Field'], Json.TypeMap['Field']>;
+  BalanceChange: ProvableExtended<
     BalanceChange,
+    Value.TypeMap['BalanceChange'],
     {
       magnitude: Json.TypeMap['UInt64'];
       sgn: Json.TypeMap['Sign'];
     }
   >;
-  Events: Signable<
+  Events: ProvableExtended<
     {
       data: Field[][];
       hash: Field;
     },
+    {
+      data: Value.TypeMap['Field'][][];
+      hash: Value.TypeMap['Field'];
+    },
     Json.TypeMap['Field'][][]
   >;
-  Actions: Signable<
+  Actions: ProvableExtended<
     {
       data: Field[][];
       hash: Field;
     },
+    {
+      data: Value.TypeMap['Field'][][];
+      hash: Value.TypeMap['Field'];
+    },
     Json.TypeMap['Field'][][]
   >;
-  ActionState: Signable<Field, Json.TypeMap['Field']>;
-  MayUseToken: Signable<
+  ActionState: ProvableExtended<Field, Value.TypeMap['Field'], Json.TypeMap['Field']>;
+  MayUseToken: ProvableExtended<
     {
       parentsOwnToken: Bool;
       inheritFromParent: Bool;
+    },
+    {
+      parentsOwnToken: Value.TypeMap['Bool'];
+      inheritFromParent: Value.TypeMap['Bool'];
     },
     {
       parentsOwnToken: Json.TypeMap['Bool'];
       inheritFromParent: Json.TypeMap['Bool'];
     }
   >;
-  VerificationKeyHash: Signable<Field, Json.TypeMap['Field']>;
-  ReceiptChainHash: Signable<Field, Json.TypeMap['Field']>;
+  VerificationKeyHash: ProvableExtended<Field, Value.TypeMap['Field'], Json.TypeMap['Field']>;
+  ReceiptChainHash: ProvableExtended<Field, Value.TypeMap['Field'], Json.TypeMap['Field']>;
 };
 let customTypes: CustomTypes = {
   TransactionVersion,
@@ -125,10 +147,11 @@ let customTypes: CustomTypes = {
   VerificationKeyHash,
   ReceiptChainHash,
 };
-let { signableFromLayout, toJSONEssential, empty } = SignableFromLayout<TypeMap, Json.TypeMap>(
+let { provableFromLayout, toJSONEssential, empty } = ProvableFromLayout<
   TypeMap,
-  customTypes
-);
+  Value.TypeMap,
+  Json.TypeMap
+>(TypeMap, customTypes);
 
 type ZkappCommand = {
   feePayer: {
@@ -339,7 +362,7 @@ type ZkappCommand = {
   memo: string;
 };
 
-let ZkappCommand = signableFromLayout<ZkappCommand, Json.ZkappCommand>(
+let ZkappCommand = provableFromLayout<ZkappCommand, Value.ZkappCommand, Json.ZkappCommand>(
   jsLayout.ZkappCommand as any
 );
 
@@ -540,7 +563,7 @@ type AccountUpdate = {
   };
 };
 
-let AccountUpdate = signableFromLayout<AccountUpdate, Json.AccountUpdate>(
+let AccountUpdate = provableFromLayout<AccountUpdate, Value.AccountUpdate, Json.AccountUpdate>(
   jsLayout.AccountUpdate as any
 );
 
@@ -593,4 +616,4 @@ type Account = {
   };
 };
 
-let Account = signableFromLayout<Account, Json.Account>(jsLayout.Account as any);
+let Account = provableFromLayout<Account, Value.Account, Json.Account>(jsLayout.Account as any);
