+++ mina-transaction/v1/types.ts
@@ -1,3 +1,3 @@
-export { provableFromLayout, toJSONEssential } from '../gen/transaction.js';
-export * as Types from '../gen/transaction.js';
-export * as TypesBigint from '../gen/transaction-bigint.js';
+export { provableFromLayout, toJSONEssential } from '../gen/v1/transaction.js';
+export * as Types from '../gen/v1/transaction.js';
+export * as TypesBigint from '../gen/v1/transaction-bigint.js';
